+++ service/test/agama/storage/config_conversions/from_model_conversions/context.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+# Copyright (c) [2026] SUSE LLC
+#
+# All Rights Reserved.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of version 2 of the GNU General Public License as published
+# by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE LLC about this file by physical or electronic mail, you may
+# find current contact information at www.suse.com.
+
+require_relative "../../storage_helpers"
+require_relative "../../product_config_context"
+require "agama/storage/system"
+require "y2storage/encryption_method"
+
+shared_context "from model conversions" do
+  include Agama::RSpec::StorageHelpers
+
+  include_context "product config"
+
+  before do
+    mock_storage(devicegraph: scenario)
+
+    # Speed up tests by avoding real check of TPM presence.
+    allow(Y2Storage::EncryptionMethod::TPM_FDE).to receive(:possible?).and_return(true)
+  end
+
+  let(:scenario) { "disks.yaml" }
+
+  let(:storage_system) { Agama::Storage::System.new }
+end
